drf-auth/party/views.py
from django.http import HttpResponse
from django.shortcuts import render
from rest_framework.generics import (
                                        ListCreateAPIView
                                    )
from rest_framework.permissions import IsAuthenticated

from .models import Party
from .serializers import PartySerializer


class PartyListView(ListCreateAPIView):
    queryset = Party.objects.all()
    serializer_class = PartySerializer
    permission_classes = (IsAuthenticated,)
